tsify ChildRowDeleteButton tests
